Move exceptions to their own file in preparation to avoid circular imports
