+++ phaser/__init__.py
@@ -27,8 +27,8 @@ have datatype casting and name canonicalization done automatically.
 # commitment to build for that.
 
 
-from phaser.pipeline import (Pipeline, PhaserError,
-                             DataErrorException, DataException, DropRowException, WarningException)
+from phaser.pipeline import Pipeline
+from phaser.exceptions import PhaserError, DataErrorException, DataException, DropRowException, WarningException
 from phaser.phase import Phase, ReshapePhase, DataFramePhase
 from phaser.steps import row_step, batch_step, context_step, check_unique, sort_by
 from phaser.column import Column, IntColumn, DateColumn, DateTimeColumn, FloatColumn
